split KeyValuePair
